se implemento los servicios para obtener la couta de credito

+++ lib/modules/credits/GetLoanFlowAnnuitiesDetailDataForCredit/domain/entities/get_loan_flow_annuities_detail_data_for_credit_entity.dart
@@ -20,7 +20,7 @@ class GetLoanFlowAnnuitiesDetailDataForCreditEntity {
     final int idCreditState;
     final String creditState;
     final int delayDays;
-    final int creditAmount;
+    final double creditAmount;
     final List<ColAnnuitiesDetail> colAnnuitiesDetail;
     final String customerName;
     final String identityCard;
